+++ lib/sim_card_code.dart
@@ -1,7 +1,6 @@
 library;
 
 import 'dart:async';
-import 'dart:developer';
 import 'package:flutter/services.dart';
 
 part 'src/core.dart';